Turn off corrects-path-p by default
So, the situation is this: all the examples in the RFC use quotes around the path part,
but the syntax description itself does not contain the quotes and IE gets confused by the
extra (?) quotes, while FF works right with and without them.

I think the entire corrects-path-p should be removed and the tests updated. 

#48